Fix: Fixed bugs in refreshControl and moved Home functions to a new component in hooks/home_functions
